Reduce usage of unsigned types (#247)

+++ include/dlaf/matrix/matrix_view_const.tpp
@@ -23,13 +23,13 @@ MatrixView<const T, device>::MatrixView(blas::Uplo uplo, MatrixType<T2, device>&
 template <class T, Device device>
 hpx::shared_future<Tile<const T, device>> MatrixView<const T, device>::read(
     const LocalTileIndex& index) noexcept {
-  std::size_t i = tileLinearIndex(index);
+  std::size_t i = static_cast<std::size_t>(tileLinearIndex(index));
   return tile_shared_futures_[i];
 }
 
 template <class T, Device device>
 void MatrixView<const T, device>::done(const LocalTileIndex& index) noexcept {
-  std::size_t i = tileLinearIndex(index);
+  std::size_t i = static_cast<std::size_t>(tileLinearIndex(index));
   tile_shared_futures_[i] = {};
 }
